spin out exemplar class into own file
